facturacion-pos-app/requirements.txt
SQLAlchemy>=2.0
pandas
openpyxl
tkinter
pyodbc
reportlab
python-escpos